Moved C++ module.
